+++ src/Shipment/Models/ShippingMethod.php
@@ -17,6 +17,7 @@ namespace Vanilo\Shipment\Models;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use Vanilo\Shipment\Traits\BelongsToCarrier;
+use Vanilo\Support\Traits\ConfigurableModel;
 
 /**
  * @property int $id
@@ -33,6 +34,7 @@ use Vanilo\Shipment\Traits\BelongsToCarrier;
 class ShippingMethod extends Model
 {
     use BelongsToCarrier;
+    use ConfigurableModel;
 
     protected $guarded = ['id', 'created_at', 'updated_at'];
 
@@ -40,17 +42,6 @@ class ShippingMethod extends Model
         'configuration' => 'json',
         'is_active' => 'bool',
     ];
-
-    public static function boot()
-    {
-        parent::boot();
-
-        static::saving(function ($model) {
-            if (null === $model->configuration) {
-                $model->configuration = [];
-            }
-        });
-    }
 
     public function scopeActives(Builder $query): Builder
     {